requirements.txt
streamlit==1.31.0
google-generativeai==0.8.6